refactor: 가독성을 위한 명칭 변경 및 파일 분리
